#4 Corregidos metodos de creacion de piso y techo.

+++ TGC.Group.Interfaces/IScenarioCreator.cs
@@ -5,6 +5,6 @@ namespace TGC.Group.Interfaces
 {
     public interface IScenarioCreator
     {
-        List<List<TgcPlane>> CreateScenario(string mediaDir, IVector3Factory vector3Factory, ITgcPlaneFactory tgcPlaneFactory, int planeSize);
+        List<List<TgcPlane>> CreateScenario(string mediaDir, IVector3Factory vector3Factory, ITgcPlaneFactory tgcPlaneFactory, float planeSize);
     }
 }
